LayoutTests/http/tests/media/media-source/video-media-source-append-in-ended-state-expected.txt

Tests calling append() after the MediaSource has transitioned to the "ended" state.

Appending initialization segment.
EVENT(loadedmetadata)
Appending first media segment.
Calling endOfStream().
Appending second media segment.
END OF TEST
